Refactor SeiunActivity
